refactor(app): move images local resources into their own dedicated namespace
