+++ app/src/local-resources/images/hooks/useImageGalleryData.ts
