+++ app/src/organisms/ODD/RunningProtocol/ImageGalleryList/__tests__/GalleryListItem.test.tsx
@@ -4,7 +4,7 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
-import { useImageGalleryData } from '/app/local-resources/dataFiles/hooks/useImageGalleryData'
+import { useImageGalleryData } from '/app/local-resources/images/hooks/useImageGalleryData'
 import { handleCameraPhotoModal } from '/app/organisms/ODD/RunningProtocol/ImageGalleryList/CameraPhotoModal'
 import { useImage } from '/app/resources/dataFiles/useImage'
 
@@ -17,7 +17,7 @@ vi.mock(
     handleCameraPhotoModal: vi.fn(),
   })
 )
-vi.mock('/app/local-resources/dataFiles/hooks/useImageGalleryData')
+vi.mock('/app/local-resources/images/hooks/useImageGalleryData')
 vi.mock('/app/resources/dataFiles/useImage')
 
 const MOCK_IMG_PATH = '/path/to/test-image.jpg'
@@ -51,6 +51,7 @@ describe('GalleryListItem', () => {
   beforeEach(() => {
     vi.mocked(useImage).mockReturnValue(MOCK_IMG_PATH)
     vi.mocked(useImageGalleryData).mockReturnValue({
+      currentCommand: {} as any,
       currentCommandString: MOCK_CMD_TEXT,
       previousCommandString: MOCK_PREV_CMD_TEXT,
       stubStepFraction: MOCK_STEP,
